+++ lib/internal/bootstrap_node.js
@@ -11,6 +11,10 @@
   let internalBinding;
   let isMainThread;
 
+  // Add process to global first as it may be used by native modules loaded
+  // later on.
+  global.process = process;
+
   function startup() {
     const EventEmitter = NativeModule.require('events');
     process._eventsCount = 0;
@@ -265,7 +269,6 @@
       enumerable: false,
       configurable: true
     });
-    global.process = process;
     const util = NativeModule.require('util');
 
     function makeGetter(name) {
