+++ node.gyp
@@ -206,6 +206,7 @@
         'src/node_constants.cc',
         'src/node_contextify.cc',
         'src/node_debug_options.cc',
+        'src/node_extras.cc',
         'src/node_file.cc',
         'src/node_http2.cc',
         'src/node_http_parser.cc',
